Split config Python file in 2 for simpler customization.
